initial first commit

+++ src/main/java/casestudy/database/DAO/CompanyDAO.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompanyDAO extends JpaRepository<Company, Long> {
+public interface CompanyDAO extends JpaRepository<Company, Integer> {
 }
