+++ src/main/java/casestudy/database/DAO/ReviewDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReviewDAO extends JpaRepository<Review, Long> {
+public interface ReviewDAO extends JpaRepository<Review, Integer> {
 
 
     //have a list of the companies and let the user review them accordingly to what they got from them
